backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database & Supabase
supabase==2.7.2
pgvector==0.3.4
asyncpg==0.29.0
sqlalchemy==2.0.35

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI & RAG
openai==1.50.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
tiktoken==0.7.0

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0

# Voice
elevenlabs==1.9.0
python-socketio==5.11.4

# Config & Utils
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Dev
httpx==0.27.2
pytest==8.3.3
